Rename and fixes
